requirements.txt
simplejson
Flask
flask-cors
gunicorn==19.7.1
mysql-connector-python